Task 5 formatting still has issues
It can be tough when the markdown iterpretation locally differs from
the GitHub or repository interpretation.  Might be better to use
the GitHub repository editor and preview if file is mainly intended
as GitHub documentation?  In any case, I think that this formatting
shows up as sub enumerates now both for md to pdf and for the
GitHub markdown rendering.